feat: support direct variable offsets up to 32-bits

+++ include/pico/binding/address_env.h
@@ -36,7 +36,7 @@ typedef struct {
     AddressEntry_t type;
     union {
         void* value;
-        int16_t stack_offset;
+        int32_t stack_offset;
     };
 } AddressEntry;
